+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+dev:
+    - DB service little bit refactored.
+      added methods:
+        - dump_db: wrapper around ``db.dump`` server method. automaticaly performs base64 decoding
+        - restore_db: wrapper around ``db.restore`` server methods. automaticaly base64 encode file before sending
+
 0.6:
     - *Backward incompatible*: Changed session file format.
       *Start up imports* and *extra_paths* moved to *options* section of file.